+++ CHANGELOG.md
@@ -2,6 +2,6 @@
 All notable changes to the ObjectStore API and assets.
 
 ## 2026-03-11
-### Fixed
-- Fix 73 broken sprite paths (case sensitivity) + add frame metadata for 1119 sprites (261217e)
+### Changed
+- Import 61 3D models, 9 textures, 6 world snapshots, 13 terrain files from WebClient (526e1ce)